api/requirements.txt
# CleanSky LA API - Requirements
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
xarray==2024.7.0
numpy==2.1.2
pandas==2.2.3
h5netcdf==1.3.0
netcdf4==1.7.2
scipy==1.13.1
python-multipart==0.0.12
Pillow==10.4.0